% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binman_utils.R
\name{app_dir}
\alias{app_dir}
\title{Get application directory}
\usage{
app_dir(appname, check = TRUE)
}
\arguments{
\item{appname}{A character string giving the name of the application}

\item{check}{check whether the app given by appname exists or not.}
}
\value{
A character string giving the path of the directory
}
\description{
Get application directory
}
\examples{
\dontrun{
appdir <- app_dir("superduperapp", FALSE)
}
}
