% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_directory.R
\name{assign_directory}
\alias{assign_directory}
\title{Assign directory}
\usage{
assign_directory(dllist, appname)
}
\arguments{
\item{dllist}{A named list of data.frames. The name indicates the
platform. The data.frame should contain the version, url and file
to be processed.}

\item{appname}{Name to give the app}
}
\value{
A named list of data.frames. The data.frame should contain the
    version, url and file to be processed, the directory to download
    the file to and whether the file already exists.
}
\description{
Assign directory to download list
}
\examples{
\dontrun{
tdata <- system.file("testdata", "test_dllist.Rdata", package="binman")
load(tdata)
assign_directory(test_dllist, "myapp")
}
}
