% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binman-package.r
\docType{package}
\name{binman}
\alias{binman}
\title{binman}
\description{
A Binary Download Manager.
}
\details{
Tools and functions for managing the download of binary files.
Binary repositories are defined in 'YAML' format. Defining new pre-download,
download and post-download templates allow additional repositories to be
added.
}
