% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fromJSON_url}
\alias{fromJSON_url}
\title{Create a connection to a url and read from JSON to an R object}
\usage{
fromJSON_url(url, .curl_options = getOption("httr_config", list(options =
  list()))[["options"]])
}
\arguments{
\item{url}{string for URL}

\item{.curl_options}{A named list of options - defaults httr config settings}
}
\description{
A wrapper for jsonlite::fromJSON that will create a curl handle to 
which you can pass additional curl options
}
\keyword{internal}
