% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdl_templates.R
\name{noproc_dlfiles}
\alias{noproc_dlfiles}
\title{Dont post process}
\usage{
noproc_dlfiles(dlfiles)
}
\arguments{
\item{dlfiles}{A data.frame of files by platform and indicating
whether they were processed}
}
\value{
Returns a list of character vectors indicating files
    processed
}
\description{
Dont post process dlfiles
}
\examples{
\dontrun{
ymlfile <- system.file("exdata", "sampleapp4.yml", package="binman")
trdata <- system.file("testdata", "test_dlres.Rdata", package="binman")
load(trdata)
testthat::with_mock(
  `httr::GET` = function(...){
    test_llres
  },
  `base::dir.create` = function(...){TRUE},
  procyml <- process_yaml(ymlfile)
)
procyml
}
}
