% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{parse_version}
\alias{parse_version}
\title{Parse semantic version}
\description{
See \code{\link[semver]{parse_version}} for more details.
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{semver}{\code{\link[semver]{parse_version}}}
}}
\keyword{internal}

