% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_yaml.R
\name{process_yaml}
\alias{process_yaml}
\title{Process a yaml file}
\usage{
process_yaml(ymlfile)
}
\arguments{
\item{ymlfile}{A file in a YAML format defining the pre-download/
download and post download functions together with their arguments.}
}
\description{
Process a yaml file. The file defines the pre-download function,
    the download function and the post download function.
}

