% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reset_version}
\alias{reset_version}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{semver}{\code{\link[semver]{reset_version}}}
}}
\keyword{internal}

