% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sem_ver}
\alias{sem_ver}
\title{Semantic versioning}
\usage{
sem_ver(semver)
}
\arguments{
\item{semver}{Character vector of package versions following semantic
versioning see \url{http://semver.org/}}
}
\value{
A list of class "svlist" containing "semver" objects.
}
\description{
Semantic versioning
}
\examples{
\dontrun{
verStr <- c("10.20.30", "1.2.3-beta",
            "10.2.3-DEV-SNAPSHOT", "1.2.3-SNAPSHOT-123")
vs <- sem_ver(verStr)
max(vs)
min(vs)
sort(vs)
range(vs)
as.character(vs)
}
}

