% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{set_version}
\alias{set_version}
\title{Set semantic version}
\description{
See \code{\link[semver]{set_version}} for more details.
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{semver}{\code{\link[semver]{set_version}}}
}}
\keyword{internal}

