% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdl_templates.R
\name{unzip_dlfiles}
\alias{unzip_dlfiles}
\title{Unzip downloaded files}
\usage{
unzip_dlfiles(dlfiles)
}
\arguments{
\item{dlfiles}{A data.frame of files by platform and indicating
whether they were process}
}
\value{
indicates whether the unzips were successful
}
\description{
Unzip downloaded files. Keeps the original zip file
}
\examples{
\dontrun{
 x<-1
}
}

