% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdl_templates.R
\name{unzip_dlfiles}
\alias{unzip_dlfiles}
\title{Unzip downloaded files}
\usage{
unzip_dlfiles(dlfiles)
}
\arguments{
\item{dlfiles}{A data.frame of files by platform and indicating
whether they were process}
}
\value{
Returns a list of character vectors indicating files
    processed
}
\description{
Unzip downloaded files. Keeps the original zip file
}
\examples{
\dontrun{
ymlfile <- system.file("exdata", "sampleapp.yml", package="binman")
trdata <- system.file("testdata", "test_dlres.Rdata", package="binman")
load(trdata)
testthat::with_mock(
  `httr::GET` = function(...){
    test_llres
  },
  `base::dir.create` = function(...){TRUE},
  `utils::unzip` = function(zipfile, ...){zipfile},
  procyml <- process_yaml(ymlfile)
)
procyml
}
}

