% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdl_templates.R
\name{unzip_dlfiles}
\alias{unzip_dlfiles}
\title{Unzip downloaded files}
\usage{
unzip_dlfiles(dlfiles)
}
\arguments{
\item{dlfiles}{A named list. The names are the relevant platforms.
The list contain named logical vectors}
}
\description{
Unzip downloaded files. Keeps the original zip file
}

